# Density fitting performance optimization (re issue #1915)