--- a/pyscf/df/df_jk.py
+++ b/pyscf/df/df_jk.py
@@ -260,7 +260,8 @@ def get_jk(dfobj, dm, hermi=1, with_j=True, with_k=True, direct_scf_tol=1e-13):
 
     if not with_k:
         for eri1 in dfobj.loop():
-            vj += numpy.matmul(numpy.matmul(dmtril, eri1.T), eri1)
+            # uses numpy.matmul
+            vj += (dmtril @ eri1.T) @ eri1
 
 
     elif getattr(dm, 'mo_coeff', None) is not None:
@@ -290,8 +291,8 @@ def get_jk(dfobj, dm, hermi=1, with_j=True, with_k=True, direct_scf_tol=1e-13):
             naux, nao_pair = eri1.shape
             assert (nao_pair == nao*(nao+1)//2)
             if with_j:
-                vj += numpy.matmul(numpy.matmul(dmtril, eri1.T), eri1)
-
+                # uses numpy.matmul
+                vj += (dmtril @ eri1.T) @ eri1
             for k in range(nset):
                 nocc = orbo[k].shape[1]
                 if nocc > 0:
@@ -317,7 +318,8 @@ def get_jk(dfobj, dm, hermi=1, with_j=True, with_k=True, direct_scf_tol=1e-13):
         for eri1 in dfobj.loop(blksize):
             naux, nao_pair = eri1.shape
             if with_j:
-                vj += numpy.matmul(numpy.matmul(dmtril, eri1.T), eri1)
+                # uses numpy.matmul
+                vj += (dmtril @ eri1.T) @ eri1
 
 
             for k in range(nset):