This pull request incorporates enhancements to density fitting proposed in #1915 by @ajz34 .

- Replace unoptimized `np.einsum` with `gemm` via `numpy.matmul`
- Replace `scipy.linalg.triangular_solve` with `?trsm` in order to avoid a single-threaded array transpose and related issues surrounding Fortran vs C order
- Since with trsm, setting `overwrite_b=True` actually _does_ cause `b` to get overwritten, we have to use `map` rather than `lib.map_with_prefetch` when writing the cderi chunks to disk
- Use of `read_direct` in `_load_from_h5g` avoids an unnecessary copy and reads the data right where they need to go

The test case is the same as in the issue: [here](https://github.com/pyscf/pyscf/files/13063667/benchmark_code.py.txt)

Benchmark, 48 cores cascade lake (CPU: 2x 24-core Intel Xeon Gold 6252N)
PySCF compiled with gcc. NumPy installed from conda-forge along with Intel MKL.

1087 cGTOs
| Task | Before | After |
| --- | --- | --- |
| density fitted RHF | 87.5568 s |  31.4202 s |
| CPU time for cholesky_eri | 25.96 s | 8.41 s |
| CPU time for df vj and vk  | 31.67 s |  11.24 s |
| `df.outcore.cholesky_eri_b` | 11.8159 s |  4.2692 s |
| `df.outcore.cholesky_eri` | 22.12 s |  10.08 s |