@chillenb 2024-05-09
@sunqm I've tried to implement the buffer-swapping technique from incore as you suggested, but I am unsure about the proper buffer size. I left buflen alone and the memory usage seems to stay well under max_memory, but I have not done extensive tests. Your advice would be appreciated!