@sunqm 2024-05-16
I'll leave this PR one more week for more comments before merging. If no more comments, this optimization will be available in next release (pyscf 2.6)